+++ e2e/test/scenarios/admin/i18n/content-translation/helpers/e2e-content-translation-helpers.ts
@@ -1,4 +1,4 @@
-import type { DictionaryArray, DictionaryResponse } from "metabase/i18n/types";
+import type { DictionaryArray, DictionaryResponse } from "metabase-types/api";
 
 export const getCSVWithHeaderRow = (dictionary: DictionaryArray) => {
   const header = ["Language", "String", "Translation"];
