+++ frontend/src/metabase-types/api/content-translation.ts
@@ -1,4 +1,10 @@
-import type { DictionaryArrayRow } from "metabase/i18n/types";
+export type DictionaryArrayRow = {
+  locale: string;
+  msgid: string;
+  msgstr: string;
+};
+
+export type DictionaryArray = DictionaryArrayRow[];
 
 /** Translations retrieved from the BE have ids */
 export type RetrievedDictionaryArrayRow = DictionaryArrayRow & { id: number };
