+++ frontend/src/metabase-types/api/index.ts
@@ -9,6 +9,7 @@ export * from "./bug-report";
 export * from "./card";
 export * from "./click-behavior";
 export * from "./collection";
+export * from "./content-translation";
 export * from "./dashboard";
 export * from "./database";
 export * from "./dataset";
