+++ frontend/src/metabase/i18n/types.ts
@@ -1,11 +1,3 @@
-export type DictionaryArrayRow = {
-  locale: string;
-  msgid: string;
-  msgstr: string;
-};
-
-export type DictionaryArray = DictionaryArrayRow[];
-
 export type NonEmpty<ArrayType> = ArrayType extends (infer ItemType)[]
   ? [ItemType, ...ItemType[]]
   : never;
